refactor(taskflow): extract feature.activity
Move ActivityModel, RecordActivity, activityReducer, ACTIVITY_CAP, and
activityLoggerMiddleware (+ BOT_ACCOUNT_ID) into package
org.reduxkotlin.sample.taskflow.feature.activity. Update all import
sites (AccountStore, BoardScreen, BoardReducersTest, ActivityLoggerTest).
